Move console commands into core-commands plugin.
