+++ scripts/buildresource/resource.json
@@ -16,8 +16,6 @@
 		"plugins/core-clothing/shared/*",
 		"plugins/core-commands/client/*",
 		"plugins/core-commands/shared/*",
-		"plugins/core-console-cmds/client/*",
-		"plugins/core-console-cmds/shared/*",
 		"plugins/core-cuffs/client/*",
 		"plugins/core-cuffs/shared/*",
 		"plugins/core-dealership/client/*",
@@ -42,6 +40,7 @@
 		"plugins/core-vitals/shared/*",
 		"plugins/core-voice/client/*",
 		"plugins/core-voice/shared/*",
+		"plugins/core-weather/client/*",
 		"plugins/example-agenda/client/*",
 		"plugins/example-agenda/shared/*",
 		"plugins/job-mule-delivery/client/*",
