+++ src/core/plugins/core-commands/shared/console-commander.ts
@@ -11,7 +11,7 @@ const Commands: ConsoleCommand = {};
  * @param {...string[]} args
  * @return {*}
  */
-function handleConsoleMessage(cmdName: string, ...args: string[]) {
+function handleConsoleMessage(cmdName: string, ...args: string[]): any {
     if (!Commands[cmdName]) {
         console.warn(`Command: ${cmdName} does not exist!`);
         return;
